Add `async_key_value_source::get_complete_results()`
Sometimes there's no good place to call `lookup()` again on the async
data source -- e.g. the container detection engine may never be called
again for a particular container (if the only process in that container
never calls `execve()` or `chroot()` or `clone()`).

The best solution in that case is to supply a callback to be ran
from the async lookup, but that introduces thread safety issues
to the involved data.

`get_complete_results()` allows batch processing of lookup results
in the main thread.