Get CRI resource limits from cgroups
If the CRI runtime is _not_ containerd, we need to find the resource
limit values ourselves directly from cgroups. However, we cannot
do this synchronously in `cri::resolve()`, because at this point
the values may not yet be set completely. Instead, we delay the cgroup
lookup by one second.

Note: we do try to get the limits synchronously as well and if we
      succeed for all the values we expect (memory limit, cpu shares,
      cpu quota and period), we don't do the second delayed lookup.

Note: success is defined as having values both larger than zero
      and smaller than an arbitrary "too large" value (4 TiB, to avoid
      overflow when converting to a 32-bit number of kilobytes)